Uber Major Update - Loop
Turned the function into a loop to reduce amount of redundant code and makes it easier to troubleshoot.
Because the TidyVerse is being used, quosures need to be specified to select specific columns in tibbles within dplyr functions.
the switch function at the begining of the loop does it for each sex at the beginning.

Added Check within so that it can still produce e.g. Females if Males fails when no tags are recovered in one of the recovery strata.

Also produces notification that output was created as it is created.